requirements.txt
Flask==2.0.1
google-cloud-datastore==2.2.0
requests